requirements.txt
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
botocore==1.34.34
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
duckdb==0.10.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gspread==5.12.4
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.3.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
natsort==8.4.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==1.5.3
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
s3fs==2024.2.0
six==1.16.0
smmap==5.0.1
sql_metadata==2.10.0
sqlparse==0.4.4
st-gsheets-connection==0.0.4
streamlit==1.31.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
